+++ Sources/Sentry/SentryEnvelope.m
@@ -163,11 +163,53 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (_Nullable instancetype)initWithAttachment:(SentryAttachment *)attachment
+                           maxAttachmentSize:(NSUInteger)maxAttachmentSize
 {
+    NSUInteger maxAttachmentSizeInBytes = 1024 * 1024 * maxAttachmentSize;
+
     NSData *data = nil;
     if (nil != attachment.data) {
+        if (attachment.data.length > maxAttachmentSizeInBytes) {
+            NSString *message =
+                [NSString stringWithFormat:@"Dropping attachment with filename '%@', because the "
+                                           @"size of the passed data with %lu bytes is bigger than "
+                                           @"the maximum allowed attachment size of %lu bytes.",
+                          attachment.filename, (unsigned long)attachment.data.length,
+                          (unsigned long)maxAttachmentSizeInBytes];
+            [SentryLog logWithMessage:message andLevel:kSentryLogLevelDebug];
+
+            return nil;
+        }
+
         data = attachment.data;
     } else if (nil != attachment.path) {
+
+        NSError *error = nil;
+        NSFileManager *fileManager = [NSFileManager defaultManager];
+        NSDictionary<NSFileAttributeKey, id> *attr =
+            [fileManager attributesOfItemAtPath:attachment.path error:&error];
+
+        if (nil != error) {
+            NSString *message = [NSString
+                stringWithFormat:@"Couldn't check file size of attachment with path: %@. Error: %@",
+                attachment.path, error.localizedDescription];
+            [SentryLog logWithMessage:message andLevel:kSentryLogLevelError];
+
+            return nil;
+        }
+
+        unsigned long long fileSize = [attr fileSize];
+
+        if (fileSize > maxAttachmentSizeInBytes) {
+            NSString *message = [NSString
+                stringWithFormat:
+                    @"Dropping attachment, because the size of the it located at '%@' with %llu "
+                    @"bytes is bigger than the maximum allowed attachment size of %lu bytes.",
+                attachment.path, fileSize, (unsigned long)maxAttachmentSizeInBytes];
+            [SentryLog logWithMessage:message andLevel:kSentryLogLevelDebug];
+            return nil;
+        }
+
         data = [[NSFileManager defaultManager] contentsAtPath:attachment.path];
     }