+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat: Add maxAttachmentSize to SentryOptions #887
 - ref: Remove SentryAttachment.isEqual and hash #885
 - ref: Remove SentryScope.isEqual and hash #884